refactor(deploy): extract shell helpers into rhoai_in_kind package
Move sh / create_resource / wait_for_webhook_service_endpoint /
gha_log_group / TestFrame out of components/deploy.py into
src/rhoai_in_kind so they can be reused and unit-tested.

deploy.py prepends ../src to sys.path so it imports the package under a
plain interpreter (CI: `python components/deploy.py`) as well as under an
editable install (local uv venv). Also parameterize the redhat-ods and
rhods-notebooks namespaces via constants.

pyproject.toml: add the uv_build build-system (makes rhoai_in_kind an
installable package) and boto3, used by create_buckets().
